Refactor: rename lib files from kebab-case to camelCase
Renamed 4 lib files to follow JavaScript/React camelCase convention:
- http-status.js → httpStatus.js
- network-monitor.js → networkMonitor.js
- network-monitor.test.js → networkMonitor.test.js
- query-client.js → queryClient.js

Updated all import statements across 5 files:
- src/App.jsx
- src/services/backend/api.js
- src/services/backend/api.test.js
- src/lib/queryClient.js
- src/pages/errors/Network/index.jsx

All tests passing (217/217).
