+++ src/lib/networkMonitor.js
