+++ src/lib/queryClient.js
@@ -1,7 +1,7 @@
 import { MutationCache, QueryCache, QueryClient } from '@tanstack/react-query'
 
-import { HttpStatus } from '@/lib/http-status'
-import { getNetworkErrorType, isNetworkError } from '@/lib/network-monitor'
+import { HttpStatus } from '@/lib/httpStatus'
+import { getNetworkErrorType, isNetworkError } from '@/lib/networkMonitor'
 import { withQuery } from '@/lib/url'
 
 const getRetryDelay = attemptIndex => Math.min(1000 * 2 ** attemptIndex, 30000) // max 30 seconds
